// abrpoint.client/src/models/Lcategorie.ts
export interface Lcategorie {
  ordre?: number;
  soccod?: string;
  catcod?: string;
  codposte?: string;
  catdu?: Date | null;
  catau?: Date | null;
  catfixe?: string;
  catlib?: string;
  cathsup?: string;
  catperiode?: string;
  catsem2?: string | null;
  catsem3?: string | null;
  catsem4?: string | null;
  catsem5?: string | null;
  catsem6?: string | null;
}